test: remove jest script
