+++ package.json
@@ -67,21 +67,17 @@
     "@rspack/cli": "workspace:*",
     "@taplo/cli": "^0.7.0",
     "@types/is-ci": "^3.0.4",
-    "@types/jest": "29.5.14",
     "@types/node": "^20.19.24",
     "check-dependency-version-consistency": "^5.0.1",
     "commander": "13.1.0",
     "cross-env": "^10.1.0",
     "husky": "^9.1.7",
     "is-ci": "4.1.0",
-    "@rstest/core": "^0.6.3",
-    "jest": "29.7.0",
-    "jest-cli": "29.7.0",
+    "@rstest/core": "^0.6.4",
     "lint-staged": "^16.2.6",
     "prettier": "3.6.2",
     "prettier-2": "npm:prettier@2.8.8",
     "rimraf": "^5.0.10",
-    "ts-jest": "29.4.5",
     "typescript": "^5.9.3",
     "webpack": "5.102.1",
     "zx": "8.8.5"
