+++ packages/rspack-test-tools/package.json
@@ -15,8 +15,7 @@
     "./package.json": "./package.json",
     "./helper/*": "./dist/helper/*.js",
     "./helper/legacy/*": "./dist/helper/legacy/*.js",
-    "./helper/util/*": "./dist/helper/util/*.js",
-    "./jest/*": "./dist/jest/*.js"
+    "./helper/util/*": "./dist/helper/util/*.js"
   },
   "scripts": {
     "build": "tsc -b ./tsconfig.build.json",
@@ -24,7 +23,7 @@
     "api-extractor": "api-extractor run --verbose",
     "api-extractor:ci": "api-extractor run --verbose || diff temp/test-tools.api.md etc/test-tools.api.md"
   },
-  "files": ["client", "dist", "template", "jest.d.ts"],
+  "files": ["client", "dist", "template", "rstest.d.ts"],
   "publishConfig": {
     "access": "public",
     "provenance": true
@@ -41,7 +40,6 @@
     "@babel/parser": "7.28.5",
     "@babel/traverse": "7.28.5",
     "@babel/types": "7.28.5",
-    "@jest/reporters": "29.7.0",
     "cross-env": "^10.1.0",
     "filenamify": "4.3.0",
     "fs-extra": "^11.3.2",
@@ -49,7 +47,6 @@
     "javascript-stringify": "^2.1.0",
     "jest-diff": "^29.7.0",
     "jest-snapshot": "29.7.0",
-    "jest-environment-node": "29.7.0",
     "jsdom": "^26.1.0",
     "memfs": "4.48.1",
     "path-serializer": "0.5.1",
