+++ packages/rspack-test-tools/src/helper/legacy/createLazyTestEnv.js
@@ -118,9 +118,8 @@ module.exports = (globalTimeout = 2000, nameSuffix = "") => {
 				);
 			});
 		},
+		/** @type {import('@rstest/core').Expect} */
 		expect,
-		jest,
-		// Compatible with rstest
-		rstest: global.rstest || global.jest
+		rstest,
 	};
 };
