+++ packages/rspack-test-tools/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "src"
+    "rootDir": "src",
+    "types": ["./rstest.d.ts"]
   },
   "include": ["src"],
   "exclude": ["viewer"],
