+++ tests/rspack-test/package.json
@@ -19,7 +19,7 @@
     "@rspack/plugin-preact-refresh": "1.1.4",
     "@rspack/plugin-react-refresh": "^1.5.2",
     "@rspack/test-tools": "workspace:*",
-    "@rstest/core": "^0.6.3",
+    "@rstest/core": "^0.6.4",
     "@swc/helpers": "0.5.17",
     "@swc/plugin-remove-console": "^10.0.0",
     "@types/babel__generator": "7.27.0",
