+++ tests/rspack-test/rstest.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="@rspack/test-tools/rstest" />
+
+export {};
