+++ website/docs/en/contribute/development/testing.mdx
@@ -41,7 +41,7 @@ You can run Rspack tests by running `./x test unit` or `pnpm run test:unit` at t
 You can also go to the `tests/rspack-test` folder and run `npm run test` to run test cases and add some arguments:
 
 - **When refreshing test snapshots is needed**: Add `-u`, like `npm run test -- -u`
-- **When filtering test cases is needed**: Add `-t`, like `npm run test -- -t configCases/asset` to only run test cases from the `tests/rspack-test/configCases/asset` folder. Pattern matching supports regex, see [jest](https://jestjs.io/docs/cli#--testnamepatternregex) for details.
+- **When filtering test cases is needed**: Add `-t`, like `npm run test -- -t configCases/asset` to only run test cases from the `tests/rspack-test/configCases/asset` folder. Pattern matching supports regex, see [rstest](https://rstest.rs/config/test/testNamePattern) for details.
 
 ### Running tests
 
@@ -50,7 +50,7 @@ You can run these test cases in the following ways:
 - Run `./x test unit` or `pnpm run test:unit` from the root directory.
 - Or run `npm run test` from the `tests/rspack-test` directory.
 - To update snapshots, run `npm run test -- -u` from the `tests/rspack-test` directory.
-- To pass specific jest cli arguments, run `npm run test -- {args}` from the `tests/rspack-test` directory.
+- To pass specific rstest cli arguments, run `npm run test -- {args}` from the `tests/rspack-test` directory.
 - To filter specific test cases, run `npm run test -- -t path-of-spec` from the `tests/rspack-test` directory.
   - Like `npm run test -- -t configCases/asset` to only run test cases from the `tests/rspack-test/configCases/asset` folder (config will be automatically mapped to configCases, and other folders will work in a similar way).
 - To use Rspack Wasm for running test cases, you need to additionally configure the following environment variables:
@@ -355,8 +355,8 @@ interface TDefaultsCaseConfig {
   cwd?: string; // process.cwd for generating the build configuration of the test case, default is the `tests/rspack-test` directory
   options?: (context: ITestContext) => TCompilerOptions<ECompilerType.Rspack>; // Test case build configuration
   diff: (
-    diff: jest.JestMatchers<Diff>,
-    defaults: jest.JestMatchers<TCompilerOptions<ECompilerType.Rspack>>,
+    diff: Assertion<Diff>,
+    defaults: Assertion<TCompilerOptions<ECompilerType.Rspack>>,
   ) => Promise<void>; // Differences from the default configuration
 }
